+++ sqlhelper.Rcheck/00_pkg_src/sqlhelper/tests/testthat/testfiles/test_prepare.sql
@@ -1,0 +1,13 @@
+-- qname = selectn
+SELECT {n};
+
+-- qname = select1
+-- quotesql = no
+-- execmethod = execute
+SELECT 1;
+
+-- qname = foocon
+-- conn_name = foo
+-- execmethod = spatial
+-- geometry = geom
+select 'foo';