+++ sqlhelper.Rcheck/00_pkg_src/sqlhelper/tests/testthat/testfiles/unknown_interpolate.sql
@@ -1,0 +1,2 @@
+-- interpolate = foo
+select 1