+++ tests/testthat/test-validate_configs.R
@@ -1,22 +1,26 @@
-test_that("config must be a list", {
-  expect_warning(validate_configs(list("c"="foo")),
-                 "Connection configuration for c was invalid; c will not be available")
 
-})
 
 test_that("All invalidations are captured", {
+
+  if(interactive()){
+    expect_warning(validate_configs(list("c"="foo")),
+                   "Connection configuration for c was invalid; c will not be available")
+  }
+
   fn <- testthat::test_path("testfiles",
                             "validate_configs_test.yml")
   configs <- read_configs(fn,exclusive=TRUE)
-  expect_warning(validate_configs(configs["no_connection"]),
-                 "Connection configuration for no_connection was invalid; no_connection will not be available")
-  expect_warning(validate_configs(configs["no_server_type"]),
-                 "Connection configuration for no_server_type was invalid; no_server_type will not be available")
-  expect_warning(validate_configs(configs["conn_child_is_list"]),
-                 "Connection configuration for conn_child_is_list was invalid; conn_child_is_list will not be available")
-  expect_warning(validate_configs(configs["no_server"]),
-                 "Connection configuration for no_server was invalid; no_server will not be available")
 
+  if(interactive()){
+    expect_warning(validate_configs(configs["no_connection"]),
+                   "Connection configuration for no_connection was invalid; no_connection will not be available")
+    expect_warning(validate_configs(configs["no_server_type"]),
+                   "Connection configuration for no_server_type was invalid; no_server_type will not be available")
+    expect_warning(validate_configs(configs["conn_child_is_list"]),
+                   "Connection configuration for conn_child_is_list was invalid; conn_child_is_list will not be available")
+    expect_warning(validate_configs(configs["no_server"]),
+                   "Connection configuration for no_server was invalid; no_server will not be available")
+  }
 
   expect_equal(validate_configs(configs["no_pool"])$no_pool$pool,FALSE)
   expect_equal(validate_configs(configs["pool_not_logical"])$pool_not_logical$pool,FALSE)