+++ NEWS.md
@@ -2,6 +2,14 @@
 
 # sqlhelper 0.1.3
 
+## Enhancements
+
+* added `run_queries()` and `run_files()` as aliases for `runqueries()` and `runfiles()`
+
+## Bug fixes
+
+* read/prep functions now pass dots instead of duplicating default values
+
 # sqlhelper 0.1.2
 
 ## Bug fixes